Add Flag to Not Include Pads When Normalizing Average Pooling (#4915)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/4915

https://github.com/pytorch/glow/issues/4565

## ModelLoaderSupport
Added a flag `CountIncludePads` to AvgPoolNode. Loading support for this arg is added in Caffe2, ONNX and Pytorch ModelLoader. In Caffe2 and Pytorch Loader, the default value is set to true while in ONNX Loader the default value is set to false.

Also support for writing ONNX Model is added.

## BackendSupoort
The support is added to CPU, NNPI, Interpreter backends.

## Unittest

No new unittest is added to **Caffe2ImporterTest**. But for existing test for avgpool, assertion is added to assert the flag is properly loaded.

Unittest `importAveragePool2DCountExcludePads` is added to **OnnxImporterTest**.

Assertion is added to existing avgpool unittest in **ONNXExporterTest**.

Unittest `AvgPoolCountExcludePads` and `Int8AvgPoolCountExcludePads` are added to **OperatorTest**. Test for Int8 quantization is mainly to test the correctness for CPU backend.

Unittest `gradientCheckAvgPoolCountExcludePads` is added to **GradCheckTest**.

Reviewed By: jfix71

Differential Revision: D23843075

fbshipit-source-id: 4db8af8274b53cb837c918e78c7ae2402c3bc36b