EliminateSliceConcat handles reshape (#4756)
Summary:
If the slice dimension is not the same as concat dimension, it may be possible to replace the slices with a merged slice + reshape if the slice dimension is adjacent to the concat dimension.

Documentation: Comments added in code to explain new optimization.

Fixes https://github.com/pytorch/glow/issues/4628

Pull Request resolved: https://github.com/pytorch/glow/pull/4756

Test Plan: A new test case is added to check for this optimization: `EliminateSliceConcatWithReshapeTest`.

Reviewed By: gcatron

Differential Revision: D24023745

Pulled By: jfix71

fbshipit-source-id: 0444e3366eda7ab224bb6d28ae0114ce78c049f7