Fix Casting from float16 to Integer Type (#4949)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/4949

Before the fix, when casting float16 to int32 or int64, we are calling static_cast on float16 directly. This gives a wrong result since float16 is defined using a proxy type uint16_t.

So when casting float16 to int type, we convert it to fp32 first then do static cast. Unittests are added to ensure the results are correct.

Reviewed By: jfix71

Differential Revision: D24067917

fbshipit-source-id: 9182dd5df26cdc2680b83510d46c3061b9e90b13