Change Inference pool compile failure to return fatal error.
Summary:
This Diff adds a new LOG_NNPI_INF_IF_ERROR_RETURN_FATAL_LLVMERROR macro and calls it in NNPIInferencePool when unable to create an NNPI device network. This signals up the stack that remediation is necessary.
Additionally we add a new counter glow.devices.fatal_error which is set when a fatal error is encountered while provisioning.

Still testing but wanted to get this up asap.

Reviewed By: yinghai

Differential Revision: D24149181

fbshipit-source-id: 6dd5512da1c33a651754e1c9466018fa379beb2f