Adding to_glow based comparison util. (#5004)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5004

With the broader introduction of to_glow, we want a mechanism for testing the relative accuracy of to_glow against our other execution modes. This diff introduces a new testing util which allow said comparison. Because to_glow only supports `nn.Module`s (not functions), all of the pure function tests had to include a class wrapper. In addition, the use of parameterized tests (using the parameterized module) was expanded, so that we can trend towards something that looks a bit more like property-based testing.

Reviewed By: yinghai

Differential Revision: D24199066

fbshipit-source-id: c21e62adec8b1a0877a36d20887b7bbe33e95a6d