Compare Shape Inference Result with Run Result
Summary:
## Context:
The old unit tests hardcoded expected shape and did not run the actual operator. And it was easily broken when there is change in operators. For example D24719909 changed operator signature and broke shape inference functions. We need to compare the shape inference engine result with actual run result. The issue was identified when run integration test and shape inference engine failed.

## Change
* Fixed quantized::embedding_bag_byte_rowwise_offsets (Expected input changed from 8 to 9)
* Change unit tests for a few operators to compare shape inference results with actual run results.
* Create separate functions for fb::embedding_bag_byte_rowwise_offsets and quantized::embedding_bag_byte_rowwise_offsets, since their function signatures are different now.

Reviewed By: allwu, qizzzh

Differential Revision: D25045423

fbshipit-source-id: ffa5801d7bcb33e031f180ca98d9052f82ac806b