Skip self in shape inference (#5104)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5104

When we run shape inference to generate compilation required shape, it's
possible that the top level graph may have a self argument. It's better to skip
it in the shape inference instead of hacking an inputMeta for it. There are
plenty of checks so if some nodes depending on it shape inference would fail
explicitly.
Also it's possible that we run multiple shape inference in one iteration - 1.
generate compilation shapes 2. runtime shape inference to slice tensor. The
graphs in two runs are different but if node block list is not empty then both
would skip generating output shapes, which is not desired. Therefore change it
to check if some dependency is missing when generating output shape and skip if
so.

Differential Revision: D25078002

fbshipit-source-id: 0aa6b45c11cd14437e0d2b48c47eb77f31141c42