Add Upsample Nearest 2D and 3D Support (#5149)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5149

- Extend the functionality of `ResizeNearestNode` in Glow to let it handle 1d and 3d cases. Before it handles 2d cases only.
- Remove `CreateUpsample` in Graph.cpp as it's used to handle `upsample_nearest3d` with `2x` scale and now `ResizeNearestNode` can handle this case.
- Add NNPI importer for `ResizeNearestNode`
- Add `Upsample_Nearest` unit tests for 1D, 2D and 3D cases. Each case is tested under `float`, `float16` and `int8` precision.
- Modify Onnx Importer and Writer to accommodate the changes above.

Note: `Upsample_Nearest3D_Int8` test fails on NNPI. I've created a repro and would need the help from Intel to debug this. Interestingly, to_glow a simple model contains quantized `upsample_nearest3d` is fine N417066. The difference between this model and the unit test is that the input of the unit test is `int8` while the model's input is `float` which is quantized afterwards.

Reviewed By: mjanderson09

Differential Revision: D25401684

fbshipit-source-id: e2dd56691b964e834a1edb9e6ba584ce99ecfac9