Removing preprocessor guards for flags (#5150)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5150

In cleaning up the flags, I tried to unify use of the `#ifdef GLOW_WITH_xxx` preprocessor guards. As a result, it required other code which relied on Flags.h to utilize those macro definitions (which some of them weren't). That resulted in subtle bugs. This diff does away with those guards. The only downside is that flags will be available that don't actually apply to your environment. Given that the CLI is not particularly human-oriented, I'm asserting that that's not a problem.

Reviewed By: jfix71

Differential Revision: D25415089

fbshipit-source-id: da6625608454591d85666b049a2c22af05b21225