Add static placeholder check before creating one in PyTorchModelLoader::loadGlowEmbeddingBag to avoid duplication
Summary: The previous logic creates one static PH for each GlowEmbeddingBag, since we assumed embedding lookups do not share the same weight tensor. This assumption is however not valid for DPER models. Therefore, we add a logic to check if a static PH is created under the same weight qualname: if so, using the existing one other than creating a new one

Reviewed By: qizzzh

Differential Revision: D25433318

fbshipit-source-id: ffaec4236b8e40f771a71fc0699d94cfe124bf7b