Fix unwindBroadcast
Summary: There was a bug in the unwindBroadcast helper, where it would return some input of a Node if it was not a Tile. This was used meant I was broken behavior in opts that used it, such as MinMax  -> Clip folding, where a `Reshape -> Max -> Min` was being turned into `Broadcast -> Clip` and leading to an invalid graph.

Reviewed By: gcatron

Differential Revision: D25505465

fbshipit-source-id: 30be60caa1fe15616434188dcd51710a0550c531