Fix unwindBroadcast (#5167)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5167

There was a bug in the unwindBroadcast helper, where it would return some input of a Node if it was not a Tile. This was used meant I was broken behavior in opts that used it, such as MinMax  -> Clip folding, where a `Reshape -> Max -> Min` was being turned into `Broadcast -> Clip` and leading to an invalid graph.

Reviewed By: gcatron, chenccfb

Differential Revision: D25505465

fbshipit-source-id: c29b18aacd89eceb19dc9ee5aafd2030a327d2d2