Let loadnorm accept float p without decimal digits (#5180)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5180

Current Glow Norm only supports L2 norm, so we have constraint in `loadnorm` to check if p is int and if p is 2. However, p can be a float as long as it doesn't have any digits after decimal point. In this case we should treat it as an int.

Reviewed By: jackm321

Differential Revision: D25540182

fbshipit-source-id: 85dbd2aba7435eb9fe2fe5258211e4fd5d77f38e