Optimize Function of Getting Graph Key
Summary:
## Context
This is to optimize hashing function of getting graph key. In the old implementation, the hash value of inputMetaStack was used as key of caching graph map. This could cause inefficiency, since sparseFeature input (one-dimensional tensor) should not affect output shape. This change is to add function to get BatchSize from inputs and use batchSize as key of caching map. If the inputs are not valid, we will fallback to original hash function.

## Change
* Added optimizedHash() function which will get batchSize from inputMetaStack. This function is guarded by Flag enableGraphMapOptimizedHash.
* Changed perGlowGraphInfoMap_ and perGlowGraphShapeMap_ to use hash value as key.

Reviewed By: allwu, qizzzh

Differential Revision: D25229683

fbshipit-source-id: 57b9a5736fdd7938bc76165386c6d0f910cf9474