Bug fix: Add type check for Constant assign()
Summary: Currently, if we assign a Tensor has different Type with a Constant's Payload to this Constant, no error will be throw; However, there will be a mismatching in Constant output type and its Payload type, which could cause fatal errors. We added a type check here to make sure the Constant output type is reset before assign().

Reviewed By: jackm321

Differential Revision: D25597546

fbshipit-source-id: e351c76e0aadde4ddae60889dff19b2164229166