Bug fix: Add type check for Constant assign() (#5190)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5190

Currently, if we assign a Tensor has different Type with a Constant's Payload to this Constant, no error will be throw; However, there will be a mismatching in Constant output type and its Payload type, which could cause fatal errors. We added a type check here to make sure the Constant output type is reset before assign().

Reviewed By: jackm321

Differential Revision: D25597546

fbshipit-source-id: 89f8664c118f79cfed7e7330b1ded5cf8c1d3446