Get Input Shapes of a Submodule (#5200)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5200

Add a function `get_submod_input_shapes` in torch_glow. Given the top-level model, the path to the submodule and the inputs of the top-level model, the function returns the input shapes of the submodule.

The reason why we don't return an input spec or even a compilation spec is that when `to_glow_selective` a Pytorch model (not traced yet) we also need the input to trace it. If a spec is returned we can't generate the input (maybe we can from the input spec?).

Reviewed By: jackm321, khabinov

Differential Revision: D25622298

fbshipit-source-id: fe0c9c28fd81c3ae05579b81ae9a9cce67af3e95