Adding IR Nodes to Glow to support calling external functions (#5204)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5204

This diff adds Glow nodes/instructions to enable call external functions invocations. One possible use of this capability is to pass a source code for a function/kernel. When processing this node, the backend can compile and execute the source code. This facility can also be used to pass a binary code or pointers to executable code.

Reviewed By: jackm321, opti-mix

Differential Revision: D25013669

fbshipit-source-id: 57a16f082793b57233a6f299aa14ce66bd08c60d