Extend implementation of aten::to operator in AI Compiler (Glow)
Summary:
Adds support for `aten::to` variants with different number of inputs. The following variants are supported:

- `to.dtype_layout(Tensor self, *, ScalarType? dtype=None, Layout? layout=None, Device? device=None, bool? pin_memory=None, bool non_blocking=False, bool copy=False, MemoryFormat? memory_format=None)`
- `to.device(Tensor self, Device device, ScalarType dtype, bool non_blocking=False, bool copy=False, MemoryFormat? memory_format=None)`
- `to.dtype(Tensor self, ScalarType dtype, bool non_blocking=False, bool copy=False, MemoryFormat? memory_format=None)`

The `aten::to.other()` form is not implemented as it seems that typical Python code will extract device/dtype from the other tensor and use the `aten::to.device` form (a Python check is added to ensure support).

The test case includes the following conversions in Python:

- `.to(torch.float)` which is converted to `aten::to.dtype()`
- `.to("cpu")` which is converted to `aten::to.dtype_layout()`
- `.to(torch.Tensor)` which is converted to `aten::to.device()`

Reviewed By: mikekgfb

Differential Revision: D25755879

fbshipit-source-id: de07801bf030cbf9186330172703e1d4967e0bc8