Check on tensor shape updated by onnx-define-symbol (#5219)
Summary:
We are not checking anywhere the shape of tensor created by use of onnx-define-symbol is valid or not. This PR created a check on the value of tensor shape updated by onnx-define-symbol and throws an error if dimension value of tensor shape proto is less than 0.

Documentation: None

Pull Request resolved: https://github.com/pytorch/glow/pull/5219

Test Plan: None

Reviewed By: jackm321

Differential Revision: D25847542

Pulled By: jfix71

fbshipit-source-id: 63b7c2637992f551db40fb4585669ce103708f4e