Express the logic in AllocationsInfo using the same logic as RuntimeBundle from BackendUtils (#5301)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5301

AllocationsInfo and RuntimeBundle were using essentially the same logic, but their equivalence was not guaranteed or verified, which could lead to very hard to find differences and bugs, especially for the AOT compilation using LLVMBackend-derived backends.

This commit expresses the logic in AllocationsInfo by literally re-using the same logic that is used by RuntimeBundle. This ensures that the addresses allocated by both methods to a given weight, placeholder or activation are always the same.

Reviewed By: jfix71

Differential Revision: D26193839

fbshipit-source-id: 73a4def9aa46497dd4ecb71ef56e3ce9325cb374