Make RecommendationSystemTest more configurable
Summary:
RecommendationSystemTest does not take any inputs and it generates "random" data and indices. But it currently does it in a very predictable way, because it always passes the same "seed" parameter to fillStableRandomData and fillStableRandomIndex functions.

This commit adds a new -random-seed command-line option which allows for changing the seed parameter and thus would allow for generating different data and indices if needed.

Reviewed By: gcatron

Differential Revision: D26253740

fbshipit-source-id: 9a8fd7e728ec850e40d9c3b030977d5f2369589d