Make RecommendationSystemTest more configurable (#5309)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5309

RecommendationSystemTest does not take any inputs and it generates "random" data and indices. But it currently does it in a very predictable way, because it always passes the same "seed" parameter to fillStableRandomData and fillStableRandomIndex functions.

This commit adds a new -random-seed command-line option which allows for changing the seed parameter and thus would allow for generating different data and indices if needed.

Reviewed By: gcatron

Differential Revision: D26253740

fbshipit-source-id: 0c104562024e506cbbdbaa9f90af913817850fea