Refactors TorchGlowBackend to new "preprocess" registration mechanism.
Summary:
Refactors TorchGlowBackend to new "preprocess" registration mechanism introduced in PR/51757

tl;dr;
This decouples preprocessing, which is meant as logic that takes place on-host, from the rest of the interface which is focused to (optionally for "compile") to run on target.

This is important for mobile use-cases where the bulk of compilation happens on preprocess, thus avoiding bringing unnecessary/impossible dependencies to the mobile binary.

Reviewed By: jackm321

Differential Revision: D26329457

fbshipit-source-id: a3ee006743a0a96921c7c220c3b0f29c3f4e847a