Use glow_snn_partitioning_num_cards as a hint and try harder during snn partitioning (#5314)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5314

Previous behavior is that we use `glow_snn_partitioning_num_cards` as a hard constraint when doing snn partitioning, which makes things a bit difficult. The diff here makes it that we use `glow_snn_partitioning_num_cards` only as a hint, and when snn partitioning fails due to not enough memory, we will increase the number of snn partitioning, using `glow_snn_partitioning_num_cards` as a step size, capped at maximum devices numbers. This will help in the case when user doesn't have a very good idea of how much memory embedding part of the model is going to take.

Reviewed By: jfix71

Differential Revision: D26299174

fbshipit-source-id: 9a5f7239d9925d60001e609c4be4f917470e7575