Add aten::full_like, aten::zeros_like, aten::ones_like, aten::empty_like support (#5332)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5332

Adds implementations of `aten::full_like`, `aten::zeros_like`, `aten::ones_like`, and `aten::empty_like` to `glow::PyTorchModelLoader`. For `empty_like`, the tensor is created via `createTouch`, which does not initialize the elements of the tensor. For all of the others, the tensor is created via `createSplat`, which initializes each element to the desired fill value.

Reviewed By: jackm321

Differential Revision: D26355845

fbshipit-source-id: 12fc3b99cf509a23e035afb344bba3af8138f712