Add support to the Provisioner for FX deferred weights
Summary:
Adds support to profisionFX() for deferredWeights. Also introduces a new field to the FX json serialization.
Nodes now have an optional "is_static" attribute which is used to signal that a placeholder represents a deferred weight.
Going this route means the the serializer will need to convert deferred weights into static placeholders.

Differential Revision: D26701965

fbshipit-source-id: d5773b49f60b42f993233914ce227add02ec0f1e