Set the p2p and drt flags in the cctx
Summary:
These were never getting set, leading to the resource validator not knowing the max size to check against. We also had two sets of flags for the same thing -- I moved the llvm version to Flags and marked it as deprecated. They alias to the same variable.

Also change the NNPI InferenceContext creation to return an error up the stack if it fails. Before it was hard aborting only in debug mode.

Reviewed By: gcatron

Differential Revision: D26736869

fbshipit-source-id: de2a0c23fb740d9fc4fc3ba3ce51fad889dcee7f