Add support to the Provisioner for FX deferred weights (#5400)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5400

Adds support to profisionFX() for deferredWeights. Also introduces a new field to the FX json serialization.
Nodes now have an optional "is_static" attribute which is used to signal that a placeholder represents a deferred weight.
Going this route means the the serializer will need to convert deferred weights into static placeholders.

Reviewed By: jfix71

Differential Revision: D26701965

fbshipit-source-id: 9ac670e966c188f6dc11ba196f445cdbc2043077