Make relu, gelu, swish, and log consistent with other unary ops
Summary:
These functions had names or signatures that did not match other create methods
in glow::Function. This change adds overloads that are consistent with the
other methods, allowing PyTorchModelLoader to wrap them like other unary ops.

The old overloads are left in place. Once references to them are switched to
the new overloads, they can be deleted.

Differential Revision: D26994437

fbshipit-source-id: cd5c0b492bba526d7e20981f90e45f929bc2e8e2