Make relu, gelu, swish, and log consistent with other unary ops (#5443)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5443

These functions had names or signatures that did not match other create methods
in glow::Function. This change adds overloads that are consistent with the
other methods, allowing PyTorchModelLoader to wrap them like other unary ops.

The old overloads are left in place. Once references to them are switched to
the new overloads, they can be deleted.

Reviewed By: mikekgfb

Differential Revision: D26994437

fbshipit-source-id: 73cf1ec4b6cf9da50c9b624d8b1e287ecbf39317