Add option for varying lengths across uniform distribution (#5439)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5439

Add support for `numIndicesPerBatchMin` and `numIndicesPerBatchMax`, where the num indices per batch is uniformly distributed between the two, inclusive.

Before we would specify an int for `numIndicesPerBatch`. Now you still can specify a single value in that same index in the input config, which should keep the same behavior. Alternatively you can specify a range like `1:2` to assign `numIndicesPerBatchMin` and `numIndicesPerBatchMax` respectively. Note that `1` is equivalent to `1:1`.

Differential Revision: D26972808

fbshipit-source-id: 4fdc33374203fb4336c3e31fa4a411e047d52fd3