transpose upsample_nearest input and output to maintain NHWC input to Glow
Summary:
transposing upsample input&output to follow Glow's layout requirements.

Why wasn't it an issue so far? ResizeLinear indicates it expects NHWC but it's not enforced and doesn't matter to logic. As long as the input and output do not have explicit layout it works. When a transposeNode with specific NHWC (conv output) is followed by an upsample it fails on layout verification.

Reviewed By: jackm321

Differential Revision: D26965137

fbshipit-source-id: 33a9c6837dc61b039e5a3b3fcb2da03284bcbfe6