Wrap upsample between 2 transpose layers when loading ONNX
Summary:
When loading an ONNX model with upsample, it is also needed to transpose the input and the output of the upsample node to follow Glow's NHWC layout requirements. For some reason this wrapping was deleted at some point and the loading started to fail. This commit adds it back.