Fix test failures in glow
Summary:
Glow some problems in its unit tests that are surfaced by updating googletest to the latest version.

# 1. Uninstantiated BackendExecTest

Googletest tells me that `BackendExecTest` is never instantiated, which is correct. I see that `BackendTest` and `BackendStatlessExecTest` are instantiated, but not `BackendExecTest`.

When I replace
```
INSTANTIATE_BACKEND_TEST(BackendTest);
```
with
```
INSTANTIATE_BACKEND_TEST(BackendExecTest);
```
Then I get test failures from `BackendTest_NNPI`. Then, with help from jfix71, the tests are fixed.

It seems like the `TEST_P(BackendExecTest, ...)` tests weren't running before, and now they are.

# 2. Uninstantiated InterpAndCPU

It seems like InterpAndCPU is just a child of Operator, so it's redundant, so I deleted it.

# 3. NumericsTest instantiated with no `TEST_P`

This isn't being used, so I commented it out.

# 4. Quantization instantiated with no `TEST_P`

This shouldn't be a parameterized test, since it uses `TEST`, not `TEST_P`. I removed the instantiation and made it a normal `Test`.

# 5. Uninstantiated BackendTest

This error went away after I switched out `INSTANTIATE_BACKEND_TEST(BackendTest);`

Differential Revision: D27351091

fbshipit-source-id: d2b6c61f4094654e298ec3dac19b07e8fd54c448