Fix the node input fetching in graph printing
Summary:
## Context
We have implemented a function to print out the graph structure which is used for shape inference input.

In several cases, we will meet shape inference missing errors. in that case, there is now no any information about the dependency.

This function will print out the graph in a DFS style which can help quickly identify which dependency is failed.

## This fix
Given a node, we will print out both input and output node. originally, for input, we need to use node->input, however, mis-wrote as node->output. this diff is to fix this.

Reviewed By: yinghai

Differential Revision: D27555793

fbshipit-source-id: 1203f29eb977dca23f87511da4bc83649ec96beb