saturateKDevices and enable available_devices in compile spec
Summary:
Enabling compiling a net on a specific number of devices - instead of saturating the entire host, if saturateKDevices is specified, only K devices will be used (Decision which physical device to use is made by Provisioner).
Adding available_devices to CompilationGroupSpec and using it to set_available_devices on compile flow.

Differential Revision: D26782089

fbshipit-source-id: 8b861c140f9e72bc8e290a00fd6bfd94c743ec78