Fix bug in importing optional params of NMS operator. (#5542)
Summary:
maxOutputBoxesPerClass, iouThreshold and scoreThreshold are optional parameters for NonMaxSupression Operator. Fix a bug in importing NMS operator when the optional parameters are not provided.
Documentation:

[Optional Fixes #issue]

Pull Request resolved: https://github.com/pytorch/glow/pull/5542

Test Plan: Added ONNX importer test to check if NonMaxSuppression operator can be imported successfully when optional parameters are not provided.

Reviewed By: jfix71

Differential Revision: D27810080

Pulled By: hl475

fbshipit-source-id: 4c1093db005da0d37f76fdb3aaa4fe999a78a5db