Add protos from OperatorTest to reload to catch BC issues (#5570)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5570

Verify backwards compatibility in Glow ONNX models by adding a bunch of Glow ONNX models based on the current Node specs in NodeGen.cpp.
- I serialized all of the models we have in OperatorTest, and saved them into `glow/tests/models/onnxGlowModels`
  - I didn't include constant data to ensure the ONNX files were small
- I added a new utility `ONNXModelLoader::createDummyConstants` to read a proto and create dummy constants for all initializers
- I added `TestImportGlowOnnxModels` to `OnnxImporterTest` which loads all of these models and verifies they don't crash.

When changing a node definition, this test should fail, assuming that node had coverage in OperatorTests.

Whenever new nodes are added, there should be some operator test added for it. We can use the flag `--operator-test-save-glow-onnx-for-test` to get the operator test to save the ONNX model for the test to a tmp location and print it out, so it can be copied to `glow/tests/models/onnxGlowModels`.

TODO: At that point we also need to add some BC code to cover loading the previous definition of the node, and add some versioning logic every time we modify a node definition so we know which version to try.

Differential Revision: D27844021

fbshipit-source-id: ca3272347421fa41edfc92a3971ab0593078aaf2