Build out torch_glow correct type mapping, convert int64 to int32 in torch_glow (#5575)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5575

This diff builds out the "correct type" mapping functionality in torch_glow which tracks the correct ScalarType of each Value so that when it differs from what Glow supports (for example uint8 represented as int8 in Glow or int64 represented as int32 in Glow) torch_glow can automatically handle the conversion between PyTorch and Glow for these values. This feature existed already for a couple of nodes but this diff ensures that every NodeValue in graphs loaded from PyTorch has a correct type and fills in many missing pieces. It also adds conversion from int64 to int32 for Glow and makes this the default.

The diff includes:
* Separates correct type specification from `addValueMapping`, adds the option to provide a function which computes the correct type for each nodes.
* Add helper functions for gather correct types like `makeCorrectType` and `getCorrectTypeFromInput`
* Validate correct types as each node is loaded, add `ValueMapping::verifyCorrectTypes` to check each `ValueMapping` to ensure that it has exactly 1 correct dtype for every Glow `NodeValue`
* track correct types for lists of `NodeValues` (for example from Tensor list Constants or `ListConstruct` nodes)
* Add int64->int32 conversion for all Glow inputs and int32->int64 conversion when `at::kLong ` is the correct type for a given output from Glow
* Add missing or fix broken correct type information for many nodes
* add int64 arithmetic and cumsum unittests
* add warning logs for when int64 inputs or constant are found
* Make glow interpreter embedding ops use int32 instead of int64

Reviewed By: spaugh, zrphercule

Differential Revision: D27850086

fbshipit-source-id: d729b60ba20dd5f772dc9f20ae14b3b3cb4d4c3c