Iterative algorithm for ChildMemSizeBasedScheduler (#5524)
Summary:
Iterative variant of ChildMemSizeBasedScheduler algorithm.
Large models with thousands of nodes (e.g. with unrolled LSTMs) could run out of stack with the original recursive algorithm.
Verified performance of the new algorithm on ~4000 node model, it's the same as the original one.

Documentation:
N/A

Pull Request resolved: https://github.com/pytorch/glow/pull/5524

Test Plan: N/A

Reviewed By: SameerAsal

Differential Revision: D28094574

Pulled By: opti-mix

fbshipit-source-id: e65034e64c8acfc46e96fff2816c5fbab8e9a800

Co-authored-by: Saurabh Shandilya <ksaurabh@cadence.com>