Wrap upsample between 2 transpose layers when loading ONNX (#5460)
Summary:
When loading an ONNX model with upsample, it is also needed to transpose the input and the output of the upsample node to follow Glow's NHWC layout requirements. For some reason this wrapping was deleted at some point and the loading started to fail. This commit adds it back.

Pull Request resolved: https://github.com/pytorch/glow/pull/5460

Reviewed By: jackm321

Differential Revision: D27956108

Pulled By: 842974287

fbshipit-source-id: 0629b4495d4e7db8dcef0be3253f6682c2e49add