Support ONNX serialization and deserialization of Glow tensor strides (#5660)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5660

Glow tensors have support for custom alignments, which affect the strides of those tensors. Since this feature is rarely used, it was never properly supported by the ONNX serialization and serialization machinery.

This diffs adds serialization/deserialization support for strides. To keep the change minimal and backwards compatible, strides are serialized optionally, only if they are non-standard, i.e. if they are used for custom tensor alignments.

Reviewed By: jfix71

Differential Revision: D28414262

fbshipit-source-id: f375fdce8fab81223c9a3c20703de868895620cf