Add failed to load weight test to DeferredWeightLoader
Summary:
Adds a new test where the first time the loader fails to load the weight, the provisioner cleans up an we add again with a functional loader. This verifies that the Provisioner cleans up properly in the case of a failed weight loading.
Additionally this adds a new error code RUNTIME_DEFERRED_WEIGHT_ERROR and adds a check in HostmanagerOnnxifi to only log(ERROR) and not fatal for this class of errors.

Differential Revision: D28460005

fbshipit-source-id: d147ce31019331e46640ca98f7912f31bd258053