Early crash when warmCache() fails at any error
Summary:
This makes the replayer fail early at any resource creation, environment initialization issues, etc. See the test plan for the error log with this diff. The diff also introduces a new macro `REPORT_AND_EXIT_ON_ERR` which extends `EXIT_ON_ERR` to report the error before exit.

This enforces the code to crash when the graph has been compiled and cached before, as the graph may be recompiled with a different setting.

Reviewed By: allwu

Differential Revision: D28328838

fbshipit-source-id: 41e972f77d8edaf61e1ee3b919b26042d5779c9b