Add failed to load weight test to DeferredWeightLoader (#5670)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5670

Adds a new test where the first time the loader fails to load the weight, the provisioner cleans up an we add again with a functional loader. This verifies that the Provisioner cleans up properly in the case of a failed weight loading.
Additionally this adds a new error code RUNTIME_DEFERRED_WEIGHT_ERROR and adds a check in HostmanagerOnnxifi to only log(ERROR) and not fatal for this class of errors.

Reviewed By: yinghai

Differential Revision: D28460005

fbshipit-source-id: 4077bd2280e681dfc5c3c2125854e997ae7db95c