Use better scale, offset and atol in quantized layernorm test
Summary:
We lower Layernorm in Glow on Interpreter backend. If it's quantized layernorm, we dequantize the input first so that lowered ops are in float precision. The reason we were doing it is because we can't figure out proper scale and offset for those lowered ops if we lower it in int8. At the end, we quantize the float output of lowered ops to get int8 output.

```
// The graph of the test after lowering looks like this

fp input -> quantize -> dequantize -> lowered ops -> quantize -> reshape -> dequantize

// Adjacent quant/dequant will get removed and final graph looks like this

fp input -> lowered ops -> quantize -> reshape -> dequantize
```

D28390125 introduces an optimization pass that would optionally sink reshape op, which makes the above graph a bit different.

```
// After the diff, graph looks like this

fp input -> lowered ops -> reshape
```
This essentially makes the quantize layernorm on Interpreter backend fully runs on fp precision, which is good in terms of precision and performance. But for testing purpose we need to add some changes so that it would still pass because the reference result that is produced by torchscript model is still run on int8 precision.

Differential Revision: D28800427

fbshipit-source-id: f8f64a608a20c534281bf7584d274fd333567510