Glow AOT serialization function (#5693)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5693

Context
To enable ahead-of-time model compilation on A*, we need to run Glow serialization logic in model transformation service to (1) precompile the remote_other net; (2) Save the compiled & partitioned & optimized Glow IR as an ONNX model file and the compilation settings to the XL format model. In this diff, we implement the function to trigger Glow serialization.

Reviewed By: wfanzju

Differential Revision: D28159708

fbshipit-source-id: b0a7377ca4e85b1f8c5e5672b814ab80d72234d0