Improve IRGen for InsertTensorNode and improve lowering of TileNode
Summary:
If InsertTensorNode completely fills the bigger tensor with the copies of the smaller one (which may happen e.g. when a TileNode is lowered into InsertTensorNode), then there is no need to first copy the old value of the big tensor as it will be completely overwritten anyways.

Improve the lowering of TileNode so that it does not use Splat to initialize the big tensor as this tensor will be completely overwritten by the copies of the small tensor represented by a tile content. Use a Touch node instead.

Reviewed By: jfix71

Differential Revision: D29685531

fbshipit-source-id: a4efd0f70912ef603b5affbbe9f089414c5a0daa