Clarify the order-sensitiveness of device config (#5746)
Summary:
As [noted in the comment of Partitioner.cpp](https://github.com/pytorch/glow/blob/master/lib/Partitioner/Partitioner.cpp#L273-L274), Glow partitioner always assigns a node to the first backend in the device config if it supports the node. We can state that behavior explicitly in the documentation to avoid confusion for the user of heterogeneous partitioning.

Pull Request resolved: https://github.com/pytorch/glow/pull/5746

Reviewed By: 842974287

Differential Revision: D29802643

Pulled By: jfix71

fbshipit-source-id: 37eb2436471f77ee37de06d8ffa4d7acc7ff3cdd