Don't sink Transpose below Quantize if Transpose has multiple non-uniform users (#5756)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5756

In general we likely only want to do this sink when (1) we expect the transpose could be eliminated when colliding with another transpose, or (2) when we know the perf will be improved via transposing less data.

So in this case where there are multiple users, only sink if all other users of the Transpose are also Quantize. Heuristically, if all users are also quantizes then it's likely that they can continue to be sunk further and then the opt would be beneficial. However if they're not all quantizes then some may get stuck and we may end up with multiple transposes when we only had one prior to the optimization.

Reviewed By: dskhudia, khabinov

Differential Revision: D29801636

fbshipit-source-id: 345a67e32c5f75521becf787ac329e247e5ed5cd