AcceleratedFusionGroup and Test
Summary:
The c++ bindings for the new AFG. Since the AFG is intended to be as slim as possible the bindings are pretty straight forward. We expose three methods: compileGraph, loadFunction, and runFunction that are used by the python AFG layer.
This also adds the AcceleratedFusionGroup python code, and a unit test.

Reviewed By: yuhc

Differential Revision: D30125409

fbshipit-source-id: 8dd6d9db53bc2877f4d1d6549a75c9be72d750de