Use null bias for FullyConnected when not used (#5851)
Summary:
**Summary**
In the TFLite importer, when the FullyConnected node does not have a bias operand we create one initialized with 0.
For such situations the TFLite importer was crashing.
Such an example is given in the model below (visualized with Netron):
![image](https://user-images.githubusercontent.com/29785908/137717723-dcd7934b-4ea1-4fee-aae6-cc38eb99434d.png)

**Test Plan**
Tested with a standalone TFLite model but cannot share it because it is private in our company.

Pull Request resolved: https://github.com/pytorch/glow/pull/5851

Reviewed By: 842974287

Differential Revision: D32050167

Pulled By: zrphercule

fbshipit-source-id: a120d9d2db1378191208660f098407fc6dd93920