Replace LOG(INFO) with VLOG(0-10)
Summary:
This diff initializes GLOG in mnist and recsys (repro already has it), but hasn't dealt with AFG C++ and Python tests, which will be supported in followup diffs. Enabling it in AFG C++ test is relatively simple, and we are still trying to figure out how to do that in Python.

To display the VLOGs, we need to specify the verbose level now: by setting `GLOG_v=X` or appending `--v=X`.

Reviewed By: jackm321

Differential Revision: D32816162

fbshipit-source-id: fdc65fca77a0867831d4d665580d1a2f32ecd932