Move FoldArithmeticChainUnderConvIntoBN to Fold pipeline
Summary:
This pass folds many ops into a BatchNorm op. It makes more sense/is more valid to have as part of the Fold pipeline.

This is necessary for fully fixing https://github.com/pytorch/glow/issues/5868 and https://github.com/pytorch/glow/issues/5892, because previous to this PR we were potentially creating a `BatchNormalizationNode` any time `optimize()` was called, but this isn't valid to do after lowering has occurred in the middle of the pipeline, because the backend (e.g. CPU backend) might not support `BatchNormalizationNode` as unlowered.

Differential Revision: D33593762

fbshipit-source-id: 4a87aeb05bcb11ae97e84d4dbfd3e9bb80f7d3bc