Move FoldArithmeticChainUnderConvIntoBN to Fold pipeline (#5901)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5901

This pass folds many ops into a BatchNorm op. It makes more sense/is more valid to have as part of the Fold pipeline.

This is necessary for fully fixing https://github.com/pytorch/glow/issues/5868 and https://github.com/pytorch/glow/issues/5892, because previous to this PR we were potentially creating a `BatchNormalizationNode` any time `optimize()` was called, but this isn't valid to do after lowering has occurred in the middle of the pipeline, because the backend (e.g. CPU backend) might not support `BatchNormalizationNode` as unlowered.

Reviewed By: scottxu0730

Differential Revision: D33593762

fbshipit-source-id: e142dadb9d2a257cb4fd3a2643d236fcf3bf2fe1