Fix quantization parameters. (#5904)
Summary:
This fix solved the problem for models with malformed quantization. The type of quantization is per channel.
The values for scale bias and filter scale for certain channels are zero.

Documentation:

[Optional Fixes #issue]

Pull Request resolved: https://github.com/pytorch/glow/pull/5904

Test Plan:
I test an internal model.

Please see a detailed explanation of how to fill out the fields in the relevant sections in PULL_REQUEST.md.

Reviewed By: jfix71

Differential Revision: D34226129

Pulled By: khabinov

fbshipit-source-id: 5526fd7d7b198c3aa380bf5dcb860c70cfcc330b