Refactor SLSBench/EBBench for common code
Summary:
Merge common code/logic from SLSBench and EBBench.  This is all now pushed within `CommonSLSEB.h`, where the enum type `BenchmarkType` will be used to branch among different functions.

Note 1: Offset variable masking and related code in SLSBench in the `runSLS` function is not implemented, since it requires additional branches and a few explicit types being declared (added code complexity).  If this is performant critical, additional changes might be needed here.

Note 2: An alternative to the above is to have a parent class (e.g. SLSBench) with EBBench inheriting most functions, but that will still require additional code sharing since `main` and `parseArg` are still part of the superclass.  This is not done here.

Reviewed By: caogao

Differential Revision: D35405678

fbshipit-source-id: 535c6cad7e464522524c56956eb94f9dec0eb579